Update modules to accommodate DesignDocument move
- Update databse.py import of DesignDocument
- Remove DesignDocument class from views.py
- Remove DesignDocument tests from view tests
